Feature works in testing - Release Candidate
App passed functional testing after disabling Proguard. Found Proguard
to be fatily renaming classes and members causing the app to crash
in a background thread. Ad unit displays as intended, all other
functionality works as expected. App is ready to get a version
bump and release as update on the Android store.