Added backend and endpoint projects
Added backend projects and endpoint projects.
Begin to create logic to replace the temporary backend of the On This Day app.
Created the logic to source the Song on day data, will add a backing store
local to the backend instead of using an outside data source that can be
shut down without warning